rejectTransaction() should not do anything
The assumption is that a bad transaction (hack) did not come from the
provider, and does not therefore require finalization.

If on the other hand, the verification service failed (network issues,
etc), we don't want to finalize the transaction and instead want to
be reminded later so that it can be verified again.